+++ pkg/log/http-in-request.go
@@ -7,28 +7,28 @@ import (
 )
 
 type inRequestLog struct {
-	Timestamp       time.Time           `json:"timestamp"`
-	LogLevel        logLevel            `json:"log-level"`
-	TraceID         string              `json:"trace-id"`
-	ServiceName     string              `json:"service-name"`
-	ServiceVersion  string              `json:"service-version"`
-	Hostname        string              `json:"hostname"`
-	ReqIP           string              `json:"req-ip"`
+	Timestamp       time.Time           `json:"timestamp,omitempty"`
+	LogLevel        logLevel            `json:"log-level,omitempty"`
+	TraceID         string              `json:"trace-id,omitempty"`
+	ServiceName     string              `json:"service-name,omitempty"`
+	ServiceVersion  string              `json:"service-version,omitempty"`
+	Hostname        string              `json:"hostname,omitempty"`
+	ReqIP           string              `json:"req-ip,omitempty"`
 	ReqMethod       string              `json:"req-method"`
 	ReqBody         string              `json:"req-body"`
-	ReqForm         string              `json:"req-form"`
+	ReqForm         string              `json:"req-form,omitempty"`
 	ReqURL          string              `json:"req-url"`
-	RealIP          string              `json:"real-ip"`
-	EventType       string              `json:"event-type"`
+	RealIP          string              `json:"real-ip,omitempty"`
+	EventType       string              `json:"event-type,omitempty"`
 	UserID          string              `json:"user-id,omitempty"`
-	PayloadLength   int64               `json:"payload-length"`
-	Header          map[string][]string `json:"header"`
-	ContentType     string              `json:"content-type"`
-	ContentEncoding string              `json:"content-encoding"`
+	PayloadLength   int64               `json:"payload-length,omitempty"`
+	Header          map[string][]string `json:"header,omitempty"`
+	ContentType     string              `json:"content-type,omitempty"`
+	ContentEncoding string              `json:"content-encoding,omitempty"`
 	// OAuth client ID
 	ClientID string `json:"client-id,omitempty"`
 	// TenantID is the ID of the tenant to which the log belongs to
-	TenantID string `json:"tenant-id"`
+	TenantID string `json:"tenant-id,omitempty"`
 }
 
 func (l *HTTPLogger) httpInRequest(req *http.Request) error {