+++ pkg/log/http-out-response.go
@@ -6,27 +6,27 @@ import (
 )
 
 type outResponseLog struct {
-	Timestamp       time.Time           `json:"timestamp"`
-	LogLevel        logLevel            `json:"log-level"`
-	TraceID         string              `json:"trace-id"`
-	ServiceName     string              `json:"service-name"`
-	ServiceVersion  string              `json:"service-version"`
-	Hostname        string              `json:"hostname"`
+	Timestamp       time.Time           `json:"timestamp,omitempty"`
+	LogLevel        logLevel            `json:"log-level,omitempty"`
+	TraceID         string              `json:"trace-id,omitempty"`
+	ServiceName     string              `json:"service-name,omitempty"`
+	ServiceVersion  string              `json:"service-version,omitempty"`
+	Hostname        string              `json:"hostname,omitempty"`
 	ReqMethod       string              `json:"req-method"`
 	ReqURL          string              `json:"req-url"`
-	EventType       string              `json:"event-type"`
+	EventType       string              `json:"event-type,omitempty"`
 	UserID          string              `json:"user-id,omitempty"`
 	ResponseCode    int                 `json:"response-code"`
 	ResponseBody    string              `json:"response-body"`
-	PayloadLength   int64               `json:"payload-length"`
-	Header          map[string][]string `json:"header"`
-	ContentType     string              `json:"content-type"`
-	ContentEncoding string              `json:"content-encoding"`
-	Duration        int64               `json:"roundtrip-duration"`
+	PayloadLength   int64               `json:"payload-length,omitempty"`
+	Header          map[string][]string `json:"header,omitempty"`
+	ContentType     string              `json:"content-type,omitempty"`
+	ContentEncoding string              `json:"content-encoding,omitempty"`
+	Duration        int64               `json:"roundtrip-duration,omitempty"`
 	// OAuth client ID
 	ClientID string `json:"client-id,omitempty"`
 	// TenantID is the ID of the tenant to which the log belongs to
-	TenantID string `json:"tenant-id"`
+	TenantID string `json:"tenant-id,omitempty"`
 }
 
 func (l *Logger) HttpOutResponse(